scripts/requirements.txt
requests==2.31.0
supabase==1.0.4
python-decouple==3.8